requirements.txt
Django>=6.0.2
djangorestframework>=3.15,<4.0
channels>=4.1,<5.0
daphne>=4.1,<5.0
gunicorn>=22.0,<24.0
uvicorn[standard]>=0.30,<1.0
psycopg[binary]>=3.2,<4.0
Pillow>=10.0,<12.0
flake8
